# Move contrast limits and gamma to the shader by vendoring vispy code