--- a/napari/_vispy/image.py
+++ b/napari/_vispy/image.py
@@ -1,0 +1,5 @@
+from .vendored import ImageVisual
+from vispy.scene.visuals import create_visual_node
+
+
+Image = create_visual_node(ImageVisual)