… napari/_vispy/vendored/__init__.py
… napari/_vispy/vendored/image.py
… napari/_vispy/vendored/volume.py