# Description
This PR supersedes #1056 by porting @tlambert03 newly contributed vispy code from https://github.com/vispy/vispy/pull/1844 and https://github.com/vispy/vispy/pull/1842 for Image and Volume which have merged or are close to merge respectively at vispy but are not yet released.

The PR tries to keep the vendored code well isolated though an additional line  `uniform float u_attenuation;` is present in line 88 of `napari/_vispy/vendored/volume.py` as we have yet to contribute the attenuated max intensity projection shader back to vispy.

The most exciting thing about the PR are the performance improvements, and these are particularly impressive for the 3D case

### before this PR
```
(base) czirwc1macos2701:napari nsofroniew$ asv run -b QtImageRenderingSuite --python=same
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_nsofroniew_opt_anaconda3_bin_python
[ 25.00%] ··· Running (benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_contrast--).
[ 50.00%] ··· Running (benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_gamma--).
[ 75.00%] ··· benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_contrast                                                                                                                        ok
[ 75.00%] ··· ======== =============
               param1
              -------- -------------
                 16      2.64±0.2ms
                 32      2.73±0.3ms
                 64      2.59±0.2ms
                128      2.64±0.2ms
                256     2.59±0.08ms
                512     2.40±0.07ms
                1024    2.46±0.06ms
                2048    2.42±0.07ms
                4096     2.64±0.4ms
              ======== =============

[100.00%] ··· benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_gamma                                                                                                                           ok
[100.00%] ··· ======== ============
               param1
              -------- ------------
                 16     5.47±0.7ms
                 32     5.58±0.4ms
                 64     5.22±0.3ms
                128     5.21±0.3ms
                256     5.18±0.3ms
                512     5.37±0.4ms
                1024    5.35±0.4ms
                2048    5.36±0.4ms
                4096    5.22±0.3ms
              ======== ============

(base) czirwc1macos2701:napari nsofroniew$ asv run -b QtVolumeRenderingSuite --python=same
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_nsofroniew_opt_anaconda3_bin_python
[ 25.00%] ··· Running (benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_contrast--).
[ 50.00%] ··· Running (benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_gamma--).
[ 75.00%] ··· benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_contrast                                                                                                                       ok
[ 75.00%] ··· ======== =============
               param1
              -------- -------------
                 16      7.72±0.1ms
                 32      7.99±0.1ms
                 64     11.4±0.08ms
                128      37.7±0.7ms
                256       251±4ms
                512      3.13±0.6s
              ======== =============

[100.00%] ··· benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_gamma                                                                                                                          ok
[100.00%] ··· ======== =============
               param1
              -------- -------------
                 16     5.22±0.06ms
                 32      5.37±0.2ms
                 64     6.45±0.07ms
                128     17.0±0.08ms
                256       81.6±2ms
                512       551±4ms
              ======== =============
```

### after this PR
```
(base) czirwc1macos2701:napari nsofroniew$ asv run -b QtImageRenderingSuite --python=same
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_nsofroniew_opt_anaconda3_bin_python
[ 25.00%] ··· Running (benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_contrast--).
[ 50.00%] ··· Running (benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_gamma--).
[ 75.00%] ··· benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_contrast                                                                                                                        ok
[ 75.00%] ··· ======== =============
               param1
              -------- -------------
                 16     2.32±0.02ms
                 32     2.37±0.05ms
                 64     2.37±0.03ms
                128     2.34±0.03ms
                256     2.41±0.04ms
                512     2.51±0.06ms
                1024    2.45±0.08ms
                2048    2.48±0.03ms
                4096     2.37±0.1ms
              ======== =============

[100.00%] ··· benchmark_qt_viewer_image.QtImageRenderingSuite.time_change_gamma                                                                                                                           ok
[100.00%] ··· ======== =============
               param1
              -------- -------------
                 16     2.06±0.02ms
                 32     2.08±0.05ms
                 64     2.06±0.05ms
                128     2.06±0.01ms
                256     2.06±0.02ms
                512     2.10±0.03ms
                1024    2.11±0.02ms
                2048    2.07±0.03ms
                4096    2.10±0.02ms
              ======== =============

(base) czirwc1macos2701:napari nsofroniew$ asv run -b QtVolumeRenderingSuite --python=same
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_nsofroniew_opt_anaconda3_bin_python
[ 25.00%] ··· Running (benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_contrast--).
[ 50.00%] ··· Running (benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_gamma--).
[ 75.00%] ··· benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_contrast                                                                                                                       ok
[ 75.00%] ··· ======== =============
               param1
              -------- -------------
                 16     2.45±0.07ms
                 32     2.62±0.04ms
                 64     3.75±0.09ms
                128      13.9±0.6ms
                256      77.9±0.7ms
                512       582±9ms
              ======== =============

[100.00%] ··· benchmark_qt_viewer_image.QtVolumeRenderingSuite.time_change_gamma                                                                                                                          ok
[100.00%] ··· ======== =============
               param1
              -------- -------------
                 16     2.24±0.04ms
                 32     2.28±0.08ms
                 64      3.34±0.1ms
                128      13.3±0.6ms
                256       77.3±1ms
                512       571±4ms
              ======== =============
```

One question I have for @tlambert03 is if he wants to make better use of the `contrast_limits_range` parameter in this new system - I recall those values potentially being useful, and I'm not making use of them here. Feel free to update this PR directly if that is the case.

Otherwise this is ready for review, with all tests passing locally (and all tests unchanged). 

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] New feature (non-breaking change which adds functionality)

# References
<!-- What resources, documentation, and guides were used in the creation of this PR? -->
<!-- If this is a bug-fix or otherwise resolves an issue, reference it here with "closes #(issue)" -->

# How has this been tested?
<!-- Please describe the tests that you ran to verify your changes. -->
- [x] example: all tests pass with my change

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works