@sofroniewn 2020-07-16
Thanks for the review @tlambert03! 

> At that point, if start you increasing the contrast limits range you will see much more lag (as it used to be). We can adjust that threshold with the Volume.clim_range_threshold parameter.

Yeah, I wonder @tlambert03 if we shouldn't just set this according to be `contrast_limits_range` and then update it whenever `contrast_limits_range` changes which won't be so often.

Does that make sense? If so do you want to quickly add that to this PR? I guess we don't need to - but might as well, no? I think that's how we can guarantee good performance when the clim is appropriate for the clim range.