@jni 2020-07-16
@sofroniewn I dunno, I don't like that the image can end up looking really terrible. I think I prefer @tlambert03's heuristics.

@tlambert03 medium/long term, how hard would it be to allow 16bit textures or even 32-bit float?