@jni 2020-07-16
@sofroniewn can you add a comment in the vendored code to indicate your modification? We should probably keep some kind of registry somehow that tells us what part of the code is not yet contributed upstream. In general, I prefer vendoring to be very temporary, rather than temporary forever. =P