@sofroniewn 2020-07-16
> @sofroniewn can you add a comment in the vendored code to indicate your modification?

Done

> We should probably keep some kind of registry somehow that tells us what part of the code is not yet contributed upstream.

I think this file organization along with that comment makes it more clear than before.

> In general, I prefer vendoring to be very temporary, rather than temporary forever. =P

Agreed, we can try and work on this once the big PR from vispy gets merged. We'll get there.

> @sofroniewn I dunno, I don't like that the image can end up looking really terrible. I think I prefer @tlambert03's heuristics.

That's fine. I will leave as is.

> @tlambert03 medium/long term, how hard would it be to allow 16bit textures or even 32-bit float?

I think this is a good goal, but my big concern is that it might lead to more install/ hardware incompatibility problems. Maybe that's not true in 2020 anymore, but in 2016 I think that was a concern.

I will merge once tests pass unless @tlambert03 or others have any additional comments.