@tlambert03 2020-07-16
> @tlambert03 medium/long term, how hard would it be to allow 16bit textures or even 32-bit float?

it's doable.  we'd need to fix [hard-coded 8bit texture sizes](https://github.com/vispy/vispy/blob/bb2cd96eb16dde1ecac3c70f9e572791bb5699e9/vispy/gloo/glir.py#L1635) in `vispy/gloo/glir.py`.  It's really just an issue of memory, so yes it will increase and we could probably catch errors with appropriate messages to "turn it back down to 8bit".

> if we shouldn't just set this according to be contrast_limits_range and then update it whenever contrast_limits_range changes which won't be so often.

yeah, if we pinned it to `contrast_limits_range`, then you'd end up with very posterized looking images every time someone has set their range too large, and then adjusted the slider to try to make the image look better.  (i.e. people will be surprised if the range of the slider is affecting their image display)